requirements.txt
crewai
python-dotenv
pydantic
fastmcp